Move remaining general Vg* types out adjacent to the things they need to merge with
